+++ src/Testcontainers/Clients/DockerImageOperations.cs
@@ -120,7 +120,9 @@ namespace DotNet.Testcontainers.Clients
           await this.Docker.Images.BuildImageFromDockerfileAsync(buildParameters, dockerfileStream, Array.Empty<AuthConfig>(), new Dictionary<string, string>(), this.traceProgress, ct)
             .ConfigureAwait(false);
 
-          var imageHasBeenCreated = await this.ExistsWithNameAsync(image.FullName, ct).ConfigureAwait(false);
+          var imageHasBeenCreated = await this.ExistsWithNameAsync(image.FullName, ct)
+            .ConfigureAwait(false);
+
           if (!imageHasBeenCreated)
           {
             throw new InvalidOperationException($"Docker image {image.FullName} has not been created.");