+++ tests/Testcontainers.Tests/Unit/Containers/Unix/Modules/Databases/MsSqlTestcontainerTest.cs
@@ -7,7 +7,6 @@ namespace DotNet.Testcontainers.Tests.Unit
   using DotNet.Testcontainers.Tests.Fixtures;
   using Xunit;
 
-  [Collection(nameof(Testcontainers))]
   public sealed class MsSqlTestcontainerTest : IClassFixture<MsSqlFixture>
   {
     private readonly MsSqlFixture msSqlFixture;