+++ tests/Testcontainers.Tests/Unit/Containers/Unix/Modules/MessageBrokers/RabbitMqTestcontainerTest.cs
@@ -3,7 +3,6 @@ namespace DotNet.Testcontainers.Tests.Unit
   using DotNet.Testcontainers.Tests.Fixtures;
   using Xunit;
 
-  [Collection(nameof(Testcontainers))]
   public sealed class RabbitMqTestcontainerTest : IClassFixture<RabbitMqFixture>
   {
     private readonly RabbitMqFixture rabbitMqFixture;