+++ tests/Testcontainers.Tests/Unit/Images/IgnoreFileTest.cs
@@ -4,7 +4,6 @@ namespace DotNet.Testcontainers.Tests.Unit
   using DotNet.Testcontainers.Tests.Fixtures;
   using Xunit;
 
-  [Collection(nameof(Testcontainers))]
   public sealed class IgnoreFileTest
   {
     [Theory]